+++ com_portfolio/domain/exceptions.py
@@ -1,2 +1,6 @@
-class InvalidPortfolioLabel(Exception):
+class MissingPortfolio(Exception):
+    pass
+
+
+class PortfolioAlreadyExists(Exception):
     pass