+++ com_portfolio/domain/repositories.py
@@ -1,4 +1,5 @@
 from abc import ABC, abstractmethod
+from uuid import UUID
 
 from .entities import Portfolio
 
@@ -16,6 +17,13 @@ class PortfolioRepositoryInterface(ABC):
     async def find(self, label: str) -> Portfolio:
         """
         Tries to find a portfolio owned by a User with USER_ID context var
-        and with passed label
-        if it doesn't work raises a UserHasNoPortfolio
+        and with passed label, if there is no a portfolio with the same label
+        raises a MissingPortfolioLabel
+        """
+
+    @abstractmethod
+    async def add(self, label: str) -> UUID:
+        """
+        Creates a portfolio owned by a User with USER_ID context var
+        with passed label or override exists, returns portfolio id
         """