+++ com_portfolio/presentation/api/factories.py
@@ -1,15 +1,14 @@
 from aiohttp import web
 
-from com_portfolio.application import Application, IdentityProviderInterface
+from com_portfolio.application import Application
 
-from .helpers import set_app, set_identity_provider
+from .helpers import set_app
 from .middlewares import register_middlewares
 from .views import register_views
 
 
 def create_web_app(
     app: Application,
-    identity_provider: IdentityProviderInterface,
 ) -> web.Application:
     web_app = web.Application()
 
@@ -17,6 +16,5 @@ def create_web_app(
     register_middlewares(web_app)
 
     set_app(web_app, app)
-    set_identity_provider(web_app, identity_provider)
 
     return web_app