+++ com_portfolio/presentation/api/middlewares.py
@@ -6,12 +6,11 @@ import inflection
 from aiohttp import web
 
 from com_portfolio.application import InvalidAccessToken
-from com_portfolio.context_vars import REQUEST_ID, USER_ID, context_var_set
-from com_portfolio.domain import InvalidPortfolioLabel
+from com_portfolio.context_vars import REQUEST_ID, context_var_set
+from com_portfolio.domain import MissingPortfolio, PortfolioAlreadyExists
 
 from .exceptions import InvalidAuthorizationHeader
-from .helpers import get_identity_provider
-from .responses import make_error_response, make_json_response
+from .responses import make_error_response
 from .types import Handler
 
 REQUEST_ID_HEADER: Final = "X-Request-ID"
@@ -19,7 +18,7 @@ REQUEST_ID_HEADER: Final = "X-Request-ID"
 
 def register_middlewares(web_app: web.Application) -> None:
     web_app.middlewares.append(request_id_middleware)
-    web_app.middlewares.append(authorization_middleware)
+    web_app.middlewares.append(authorization_errors_middleware)
     web_app.middlewares.append(user_portfolio_errors_middleware)
 
 
@@ -37,21 +36,16 @@ async def request_id_middleware(
 
 
 @web.middleware
-async def authorization_middleware(
+async def authorization_errors_middleware(
     request: web.Request,
     handler: Handler,
 ) -> web.StreamResponse:
-    identity_provider = get_identity_provider(request.app)
     try:
-        access_token = _get_access_token(request)
-        user_id = await identity_provider.get_user_id(access_token)
+        return await handler(request)
     except (InvalidAuthorizationHeader, InvalidAccessToken) as e:
         message = _humanize_exception(e)
         request.app.logger.info(message)
         return make_error_response(message, HTTPStatus.UNAUTHORIZED)
-    else:
-        with context_var_set(var=USER_ID, value=user_id):
-            return await handler(request)
 
 
 @web.middleware
@@ -61,20 +55,9 @@ async def user_portfolio_errors_middleware(
 ) -> web.StreamResponse:
     try:
         return await handler(request)
-    except InvalidPortfolioLabel as e:
-        response_body = {
-            "message": _humanize_exception(e),
-        }
-        return make_json_response(response_body, HTTPStatus.OK)
-
-
-def _get_access_token(request: web.Request) -> str:
-    try:
-        _, token = request.headers['Authorization'].strip().split()
-    except (KeyError, ValueError) as e:
-        raise InvalidAuthorizationHeader from e
-    else:
-        return token
+    except (MissingPortfolio, PortfolioAlreadyExists) as e:
+        message = _humanize_exception(e)
+        return make_error_response(message, HTTPStatus.BAD_REQUEST)
 
 
 def _humanize_exception(exc: Exception) -> str: